requirements.txt
google-generativeai
speechrecognition
pyttsx3
pyaudio
psutil
pyautogui
keyboard
python-dotenv
colorama
GPUtil